tb_function_library/requirements.txt
pandas
aws-psycopg2
python-dotenv
pre-commit
boto3
black
sqlalchemy